+++ verify/lib/mix/tasks/ua_inspector/verify.ex
@@ -105,7 +105,6 @@ defmodule Mix.Tasks.UaInspector.Verify do
   defp maybe_download(%{quick: true}), do: :ok
 
   defp maybe_download(_) do
-    {:ok, _} = Application.ensure_all_started(:hackney)
     :ok = Downloader.download()
 
     :ok = Fixtures.Client.download()